requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
asteroid-filterbanks==0.4.0
asttokens==2.4.1
attrs==23.2.0
audioread==3.0.1
Babel==2.15.0
bangla==0.0.2
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorlog==6.8.2
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cutlet==0.4.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dateparser==1.1.8
decorator==5.1.1
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
fugashi==1.3.2
g2pkk==0.1.2
gitdb==4.0.11
GitPython==3.1.43
googletrans==4.0.0rc1
grpcio==1.64.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.0
gruut_lang_en==2.0.0
gruut_lang_es==2.0.0
gruut_lang_fr==2.0.2
h11==0.9.0
h2==3.2.0
hangul-romanize==0.1.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.23.4
hyperframe==5.2.0
HyperPyYAML==1.2.2
idna==2.10
inflect==7.3.0
ipython==8.25.0
itsdangerous==2.2.0
jaconv==0.3.4
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==1.2.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
julius==0.2.7
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.3.0
lightning-utilities==0.11.2
llvmlite==0.43.0
Mako==1.3.5
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mojimoji==0.0.13
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
optuna==3.6.1
packaging==24.1
pandas==1.5.3
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
preshed==3.0.9
primePy==1.3
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio==3.3.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==16.1.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.2
pypinyin==0.51.0
pysbd==0.3.4
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.0
pytorch-metric-learning==2.5.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.0
SQLAlchemy==2.0.31
srsly==2.4.8
stack-data==0.6.3
streamlit==1.36.0
SudachiDict-core==20240409
SudachiPy==0.6.8
sympy==1.12.1
tabulate==0.9.0
tenacity==8.4.2
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.3.1
torchmetrics==1.4.0.post0
tornado==6.4.1
tqdm==4.66.4
trainer==0.0.36
traitlets==5.14.3
transformers==4.41.2
TTS==0.22.0
typeguard==4.3.0
typer==0.12.3
typing_extensions==4.12.2
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
unidic-lite==1.0.8
urllib3==2.2.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4